% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_trackr_dir.R
\name{clean_trackr_dir}
\alias{clean_trackr_dir}
\title{clean_trackr_dir}
\usage{
clean_trackr_dir(trackr_dir)
}
\arguments{
\item{trackr_dir}{path, path of trackr log files}
}
\description{
Remove files in the trackr_dir
}
