% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_history.R
\name{trackr_history}
\alias{trackr_history}
\title{trackr_history}
\usage{
trackr_history(trackr_id, trackr_dir, return_records = "all")
}
\arguments{
\item{trackr_id}{string trackr_id used as starting point for extracting lineage}

\item{trackr_dir}{path, path of trackr log files}

\item{return_records}{optional character, require that all log files have a corresponding data log or return any available logged records, default all}
}
\value{
list of dataframes with all parent records
}
\description{
Get the parent rows of a trackr_id
}
