% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_network.R
\name{trackr_network}
\alias{trackr_network}
\title{trackr_network}
\usage{
trackr_network(trackr_dir, ...)
}
\arguments{
\item{trackr_dir}{path, path to a directory of trackr log files}

\item{...}{arguments passed to networkD3::simpleNetwork}
}
\value{
networkD3 network graph
}
\description{
Create a simple network plot from a collectino of trackr log files. Requires networkD3.
}
