% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trackr.R
\name{trackr_new}
\alias{trackr_new}
\title{trackr_new}
\usage{
trackr_new(
  dataframe,
  trackr_dir = "~/Documents/Personal/trackr_dev/trackr_dir",
  timepoint_message = NULL
)
}
\arguments{
\item{dataframe}{data.frame, data to be added to trackr}

\item{trackr_dir}{path, path to store log files}

\item{timepoint_message}{optional character, message to identify timepoint - similar to a git commit message}
}
\value{
data.frame with added trackr_id column
}
\description{
create a new tracker file, the starting point of a tracking history
}
