% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_record_lineage.R
\name{trackr_record_lineage}
\alias{trackr_record_lineage}
\title{trackr_record_lineage}
\usage{
trackr_record_lineage(trackr_id, trackr_dir, return_class = "data.frame")
}
\arguments{
\item{trackr_id}{string trackr_id used as starting point for extracting lineage}

\item{trackr_dir}{path, path of trackr log files}

\item{return_class}{optional character, return a data.frame or list, default data.frame}
}
\value{
data.frame or list with lineage for one trackr_id
}
\description{
Get the lineage of a trackr_id
}
