% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_root.R
\name{trackr_root}
\alias{trackr_root}
\title{trackr_root}
\usage{
trackr_root(trackr_id, trackr_dir)
}
\arguments{
\item{trackr_id}{A string, the trackr_id used as a starting point for extracting the record lineage.}

\item{trackr_dir}{A string, path to trackr log files.}
}
\value{
A list, the root record for a given trackr_id.
}
\description{
Get the root record of a trackr_id.
}
\seealso{
trackr_lineage trackr_history
}
