% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_summarise.R
\name{trackr_summarise}
\alias{trackr_summarise}
\title{trackr_summarise}
\usage{
trackr_summarise(dataframe, ...)
}
\arguments{
\item{dataframe}{data.frame, data to summarised}

\item{...}{arguments passed to dplyr::summarise}
}
\value{
A summarised dataframe
}
\description{
Wrapper for dplyr::summarise to summarise data and retain trackr_ids

Summarise data with dplyr::summarise and retain trackr_ids
}
