% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_summary.R
\name{trackr_summary}
\alias{trackr_summary}
\title{trackr_summary}
\usage{
trackr_summary(lineage_file)
}
\arguments{
\item{lineage_file}{path, path to lineage file for a certain trackr_id}
}
\value{
A dataframe, a summary of a trackr lineage file.
}
\description{
Process the lineage (parent trackr_ids) of a trackr_id.
}
