% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_log.R
\name{validate_data_log}
\alias{validate_data_log}
\title{validate_data_log}
\usage{
validate_data_log(log_file)
}
\arguments{
\item{log_file}{A string, the path to a log file}
}
\value{
A boolean, whether the log file has been validated successfully (TRUE/FALSE)
}
\description{
Validate a data log file. The data log file is the hash of all row hashes.
}
