% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_data_log.R
\name{verify_data_log}
\alias{verify_data_log}
\title{verify_data_log}
\usage{
verify_data_log(log_file)
}
\arguments{
\item{log_file}{A string, the path to a log file}
}
\value{
A boolean, whether the log file has been validated successfully (TRUE/FALSE)
}
\description{
Verify a data log file. The data log file is the hash of all row hashes.
}
