% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_history.R
\name{trackr_history}
\alias{trackr_history}
\title{trackr_history}
\usage{
trackr_history(lineage_file, trackr_dir, return_records = "all")
}
\arguments{
\item{lineage_file}{A file path, path to lineage file for a certain trackr_id}

\item{trackr_dir}{A string, path to store trackr log files.}

\item{return_records}{A string, return "any" available logged records or require data logs for "all" timepoints.}
}
\value{
A list with the parent parent records for a given trackr_id.
}
\description{
Get the parent records of a trackr_id.
}
\seealso{
trackr_lineage
}
