% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_history.R
\name{trackr_history}
\alias{trackr_history}
\title{trackr_history}
\usage{
trackr_history(trackr_id, trackr_dir, return_records = "all")
}
\arguments{
\item{trackr_id}{A string, the trackr_id used as a starting point for extracting the record lineage.}

\item{trackr_dir}{A string, path to trackr log files.}

\item{return_records}{A string (optional), whether to require data log files for all processing steps "all", or any that are available "any".}
}
\value{
A list with the parent parent records for a given trackr_id.
}
\description{
Get the parent records of a trackr_id.
}
\seealso{
trackr_lineage
}
