% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_lineage.R
\name{trackr_lineage}
\alias{trackr_lineage}
\title{trackr_lineage}
\usage{
trackr_lineage(trackr_id, trackr_dir)
}
\arguments{
\item{trackr_id}{A string, the trackr_id used as a starting point for extracting the record lineage.}

\item{trackr_dir}{A string, path to trackr log files.}
}
\description{
Get the lineage (parent trackr_ids) of a trackr_id and store it in the trackr_dir.
}
