% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_network.R
\name{trackr_network}
\alias{trackr_network}
\title{trackr_network}
\usage{
trackr_network(lineage_file, ...)
}
\arguments{
\item{lineage_file}{path, path to lineage file for a certain trackr_id}

\item{...}{arguments passed to networkD3::simpleNetwork}
}
\value{
networkD3 network graph
}
\description{
Create a simple network plot for a given trackr_lineage file. Requires networkD3.
}
