% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_new.R
\name{trackr_new}
\alias{trackr_new}
\title{trackr_new}
\usage{
trackr_new(
  dataframe,
  trackr_dir = NULL,
  timepoint_message = NULL,
  log_data = FALSE
)
}
\arguments{
\item{dataframe}{data.frame, data to be added to trackr}

\item{trackr_dir}{path, path to store log files}

\item{timepoint_message}{optional character, message to identify timepoint - similar to a git commit message}

\item{log_data}{optional boolean, output a dataset log}
}
\value{
data.frame with added trackr_id column
}
\description{
create a new tracker file, the starting point of a tracking history
}
