% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_new.R
\name{trackr_new}
\alias{trackr_new}
\title{trackr_new}
\usage{
trackr_new(
  dataframe,
  trackr_dir = NULL,
  timepoint_message = NULL,
  log_data = TRUE
)
}
\arguments{
\item{dataframe}{A data.frame, the dataset to be logged}

\item{trackr_dir}{A string, path to store trackr log files.}

\item{timepoint_message}{A string (optional), a message to identify the timepoint - similar to a git commit message.}

\item{log_data}{A boolean (optional), output a full dataset log with each trackr file. Default is "TRUE"}
}
\value{
A data.frame with a trackr_id column added. Trackr log and data log files are written into the trackr_dir.
}
\description{
Create a new tracker file, the starting point of a data log history.
}
