% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_summary.R
\name{trackr_summary}
\alias{trackr_summary}
\title{trackr_summary}
\usage{
trackr_summary(trackr_dir)
}
\arguments{
\item{trackr_dir}{A string, path to trackr log files.}
}
\value{
A data.frame, a summary of all trackr log files stored in the trackr_dir.
}
\description{
Create a summary data.frame from a directory of trackr log files
}
