% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_summary.R
\name{trackr_summary}
\alias{trackr_summary}
\title{trackr_summary}
\usage{
trackr_summary(trackr_dir)
}
\arguments{
\item{trackr_dir}{path, path to a directory of trackr log files}
}
\value{
data.frame summary of trackr log files
}
\description{
Create a summary data.frame from a directory of trackr log files
}
