% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_timepoint.R
\name{trackr_timepoint}
\alias{trackr_timepoint}
\title{trackr_timepoint}
\usage{
trackr_timepoint(
  dataframe,
  trackr_dir = NULL,
  timepoint_message = NULL,
  log_data = TRUE,
  suppress_success = FALSE
)
}
\arguments{
\item{dataframe}{A data.frame, the data to be logged.}

\item{trackr_dir}{A string, path to store trackr log files.}

\item{timepoint_message}{A string (optional), a message to identify the timepoint - similar to a git commit message.}

\item{log_data}{A boolean (optional), output a full dataset log with each trackr file. Default is "TRUE"}

\item{suppress_success}{A boolean (optional), suppress success messages. Default is "FALSE".}
}
\value{
A data.frame with an updated trackr_id column. Trackr log and data log files are written into the trackr_dir.
}
\description{
Log a timepoint in the data processing chain.
}
