% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackr_timepoint.R
\name{trackr_timepoint}
\alias{trackr_timepoint}
\title{trackr_timepoint}
\usage{
trackr_timepoint(
  dataframe,
  trackr_dir = "~/Documents/Personal/trackr_dev/trackr_dir",
  timepoint_message = NULL
)
}
\arguments{
\item{dataframe}{data.frame, data to be logged}

\item{trackr_dir}{path, path of trackr log files}

\item{timepoint_message}{optional character, message to identify timepoint - similar to a git commit message}
}
\value{
data.frame with updated trackr_id column
}
\description{
Log a data processing timepoint and write a trackr file log changes
}
