% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_data_log.R
\name{verify_data_log}
\alias{verify_data_log}
\title{verify_data_log}
\usage{
verify_data_log(log_file)
}
\arguments{
\item{log_file}{character path to log file}
}
\value{
boolean, log file validated (Y/N)
}
\description{
Log a data processing timepoint
}
